added cart count fixed image size
